==> 91_appendix_simple_models_indicator.ipynb <==
## Appendix: Simple models with dummy
*Tim Braams (8460701), Vinh Phan (8462380), Maximilian Pintilie (8462780), Rahul Singh (8464147), Kartik Vijay (8463465), Diego Zucchino (8345420)*   
  
This notebook is the same as the 90_appendix notebook with simple models but with an added dummy variables for the asset type to capute crypto specific details.

import os
import pandas as pd
from sklearn.metrics import mean_squared_error, mean_absolute_error

import scripts
import scripts.ResultStore as rs
import scripts.FitPredict as fp
pd.set_option("display.precision", 8)

### Load data

train = pd.read_json("../02_Pool_Data/pooled_train_300.json")
test = pd.read_json("../02_Pool_Data/pooled_test_300.json")

train = train.loc[~(train==0).any(axis=1)]
y_train = train["returns"]
X_train = train.drop(["index", "returns", "asset"], axis=1)
X_train = pd.concat([X_train, pd.get_dummies(train["asset"])], axis=1)

y_test = test["returns"]
X_test = test.drop(["index", "returns", "asset"], axis=1)
X_test = pd.concat([X_test, pd.get_dummies(test["asset"])], axis=1)

### Helper functions

results = rs.ResultStore(load_if_exists=False)

### Baseline

from sklearn.dummy import DummyRegressor
dummy_model, dummy_predictions, dummy_results = fp.fit_predict(DummyRegressor(strategy="mean"), 
                                                            X_fit=X_train, y_fit=y_train,
                                                            X_validate=X_test, y_validate=y_test,
                                                            store=results)

### Linear Regression
https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LinearRegression.html#sklearn.linear_model.LinearRegression

from sklearn.linear_model import LinearRegression
linear_model, linear_predictions, linear_results = fp.fit_predict(LinearRegression(), 
                                                               X_fit=X_train, y_fit=y_train,
                                                               X_validate=X_test, y_validate=y_test, 
                                                               store=results)

### Support Vector Machine
https://scikit-learn.org/stable/modules/generated/sklearn.svm.SVR.html#sklearn.svm.SVR

from sklearn.svm import SVR

svr_model, svr_predictions, svr_results = fp.fit_predict(SVR(kernel="rbf", C=1, gamma=0.1, epsilon=0.1), store=results, X_fit=X_train, y_fit=y_train,
                                                               X_validate=X_test, y_validate=y_test)

### ElasticNet
https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.ElasticNet.html#sklearn.linear_model.ElasticNet

from sklearn.linear_model import ElasticNet

elastic_model, elastic_predictions, elastic_results = fp.fit_predict(ElasticNet(), 
                                                                     X_fit=X_train, y_fit=y_train,
                                                               X_validate=X_test, y_validate=y_test, store=results)

### PLS Regression
https://scikit-learn.org/stable/modules/generated/sklearn.cross_decomposition.PLSRegression.html#sklearn.cross_decomposition.PLSRegression

from sklearn.cross_decomposition import PLSRegression

pls_model, pls_predictions, pls_results = fp.fit_predict(PLSRegression(), X_fit=X_train, y_fit=y_train,
                                                               X_validate=X_test, y_validate=y_test, store=results)

### Gradient Boosting
https://scikit-learn.org/stable/modules/generated/sklearn.ensemble.GradientBoostingRegressor.html#sklearn.ensemble.GradientBoostingRegressor

from sklearn.ensemble import GradientBoostingRegressor

gbr_model, gbr_predictions, gbr_results = fp.fit_predict(GradientBoostingRegressor(), X_fit=X_train, y_fit=y_train,
                                                               X_validate=X_test, y_validate=y_test, store=results)

### XGBoost
https://xgboost.readthedocs.io/en/stable/python/python_api.html#module-xgboost.sklearn

from xgboost import XGBRegressor

xgb_model, xgb_predictions, xgb_results = fp.fit_predict(XGBRegressor(), X_fit=X_train, y_fit=y_train,
                                                               X_validate=X_test, y_validate=y_test, store=results)

### MLP
https://scikit-learn.org/stable/modules/generated/sklearn.neural_network.MLPRegressor.html#sklearn.neural_network.MLPRegressor

from sklearn.neural_network import MLPRegressor

mlp_model, mlp_predictions, mlp_results = fp.fit_predict(MLPRegressor(), X_fit=X_train, y_fit=y_train,
                                                               X_validate=X_test, y_validate=y_test, store=results)

### Results
The results are very similar or worse than for the same models without dummies. As we try to keep our data sparse, we will not use dummies for further analysis.

# results with dummy
results.get_df().sort_values("mse")

# results without dummy
rs.ResultStore(load_if_exists=True, path="results", name="simple.json").get_df().sort_values("mse")

results.save(path="results", name="simple_dummy.json")
