==> 92_appendix_eda.ipynb <==
## Appendix: Additional Plots
*Tim Braams (8460701), Vinh Phan (8462380), Maximilian Pintilie (8462780), Rahul Singh (8464147), Kartik Vijay (8463465), Diego Zucchino (8345420)* 

import numpy as np
import csv
import pandas as pd
from matplotlib import *
import sys
from pylab import *
from plotly.subplots import make_subplots
import plotly.graph_objects as go
from statsmodels.graphics import tsaplots
from statsmodels.tsa.seasonal import seasonal_decompose
from statsmodels.graphics.tsaplots import plot_pacf
import seaborn as sns
import statsmodels.tsa.api as smt
import scipy.stats as scs
from arch import arch_model
import matplotlib as mpl
import statsmodels.api as sm
import matplotlib.pyplot as plt
from pandas.plotting import lag_plot
import matplotlib.ticker as ticker
from dateutil.parser import parse
import json
%matplotlib inline

df=pd.read_json("../02_Pool_Data/pooled_train_300.json")
df['index']= pd.to_datetime(df['index'])
#Adding a year column to better categorize and group the data
df["year"] = df["index"].dt.year
df=df.sort_index()
#Removing a single row from the analysis dataset due to its extremely high return skewing all plots
df = df.loc[df["returns"]<=17]

# lag scatter plot
lag_plot(df['returns'])
plt.show()

# box plot grouped by year
sns.boxplot(data=df, x="year", y="returns")
plt.show()

# box plot grouped by year and further categorized by cryptocurrency
sns.boxplot(data=df, x="year", y="returns",hue='asset')
plt.rc("figure", figsize=(24,16))
plt.show()

#Autocorrelation Function per Cryptocurrency

crypto_list = list(df.asset.unique())
crypto_returns= dict()
for crypto in crypto_list: 
    temp_df = df[df['asset']==crypto]
    crypto_returns[crypto] =temp_df.returns.values.tolist()
    plt.rc("figure", figsize=(24,16))
    sm.graphics.tsa.plot_acf(crypto_returns[crypto], title=crypto,alpha=0.05,lags=30)
#crypto_returns['TRX']
